fix(mcp): break callLogs↔compliance ESM cycle that deadlocks MCP server on Node.js 24 (#2650)
The bundled MCP server hung indefinitely on Node.js 24 because the
esbuild output produced a circular __esm init: callLogs.ts imported
isNoLog from ../compliance, and compliance/index.ts imported
deleteCallLogsBefore/trimCallLogsToMaxRows from ../usage/callLogs.
Node 20/22 silently short-circuited the cycle; Node 24's stricter ESM
evaluation correctly waits for both inits to complete, deadlocking.

Extract the no-log opt-out state (noLogKeys, noLogDbCache, isNoLog,
setNoLog, etc.) to src/lib/compliance/noLog.ts — a leaf module that
imports only from db/core and adapter types. Switch the callers that
were closing the cycle (usage/callLogs.ts, db/detailedLogs.ts,
db/apiKeys.ts) to import directly from compliance/noLog. compliance/
index.ts re-exports isNoLog/setNoLog so external callers keep their
imports unchanged.

Add tests/unit/compliance-noLog-cycle.test.ts with a 5s race that
fails fast if the cycle regresses, plus an assertion that state
written via compliance/index is visible via compliance/noLog.

Reported in #2650.
