+++ src/lib/compliance/index.ts
@@ -413,90 +413,10 @@ export function countAuditLog(filter: AuditLogFilter = {}) {
 }
 
 // ─── No-Log Opt-Out ────────────────
-
-/** @type {Set<string>} API key IDs with logging disabled */
-const noLogKeys = new Set();
-const noLogDbCache = new Map<string, { value: boolean; timestamp: number }>();
-let noLogColumnVerified = false;
-let hasNoLogColumn = false;
-const NO_LOG_CACHE_TTL_MS = 30_000;
-const noLogIdsFromEnv = (process.env.NO_LOG_API_KEY_IDS || "")
-  .split(",")
-  .map((value) => value.trim())
-  .filter(Boolean);
-for (const id of noLogIdsFromEnv) {
-  noLogKeys.add(id);
-}
-
-/**
- * Set whether an API key opts out of request logging.
- *
- * @param {string} apiKeyId
- * @param {boolean} noLog
- */
-export function setNoLog(apiKeyId: string, noLog: boolean) {
-  if (noLog) {
-    noLogKeys.add(apiKeyId);
-  } else {
-    noLogKeys.delete(apiKeyId);
-  }
-  noLogDbCache.set(apiKeyId, { value: noLog, timestamp: Date.now() });
-}
-
-function ensureNoLogColumn(db: SqliteAdapter) {
-  if (noLogColumnVerified) {
-    return hasNoLogColumn;
-  }
-
-  try {
-    const columns = db.prepare("PRAGMA table_info(api_keys)").all() as Array<{ name: string }>;
-    hasNoLogColumn = columns.some((column) => column.name === "no_log");
-  } catch {
-    hasNoLogColumn = false;
-  }
-
-  noLogColumnVerified = true;
-  return hasNoLogColumn;
-}
-
-function readNoLogFromDb(apiKeyId: string): boolean {
-  const db = getDb();
-  if (!db || !apiKeyId) return false;
-  if (!ensureNoLogColumn(db)) return false;
-
-  const cached = noLogDbCache.get(apiKeyId);
-  if (cached && Date.now() - cached.timestamp < NO_LOG_CACHE_TTL_MS) {
-    return cached.value;
-  }
-
-  try {
-    const row = db.prepare("SELECT no_log FROM api_keys WHERE id = ?").get(apiKeyId) as
-      | { no_log?: number }
-      | undefined;
-    const value = Boolean(row && Number(row.no_log) === 1);
-    noLogDbCache.set(apiKeyId, { value, timestamp: Date.now() });
-    return value;
-  } catch {
-    return false;
-  }
-}
-
-/**
- * Check if an API key has opted out of logging.
- *
- * @param {string} apiKeyId
- * @returns {boolean}
- */
-export function isNoLog(apiKeyId: string) {
-  if (!apiKeyId) return false;
-  if (noLogKeys.has(apiKeyId)) return true;
-
-  const persistedNoLog = readNoLogFromDb(apiKeyId);
-  if (persistedNoLog) {
-    noLogKeys.add(apiKeyId);
-  }
-  return persistedNoLog;
-}
+// Moved to ./noLog.ts to break the callLogs → compliance → callLogs ESM cycle
+// that deadlocks the bundled MCP server under Node.js 24 (#2650). Re-exported
+// here so downstream callers that import from "../compliance" keep working.
+export { isNoLog, setNoLog } from "./noLog";
 
 // ─── Log Retention / Cleanup ────────────────
