+++ src/lib/db/detailedLogs.ts
@@ -8,7 +8,7 @@
 import { v4 as uuidv4 } from "uuid";
 import { getDbInstance } from "./core";
 import { getSettings } from "./settings";
-import { isNoLog } from "../compliance";
+import { isNoLog } from "../compliance/noLog";
 import {
   protectPayloadForLog,
   serializePayloadForStorage,
